requirements.txt
streamlit>=1.28.0
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0
requests>=2.31.0
Pillow>=10.0.0

